requirements.txt
Flask 1.1.2
Flask-WTForms
Flask-Login
